+++ crates/system/os/accounts/src/delete_account/support/outcome.rs
@@ -30,3 +30,22 @@ impl HasSampleValues for CreateDeleteAccountManifestOutcome {
         )
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[allow(clippy::upper_case_acronyms)]
+    type SUT = CreateDeleteAccountManifestOutcome;
+
+    #[test]
+    fn equality() {
+        assert_eq!(SUT::sample(), SUT::sample());
+        assert_eq!(SUT::sample_other(), SUT::sample_other());
+    }
+
+    #[test]
+    fn inequality() {
+        assert_ne!(SUT::sample(), SUT::sample_other());
+    }
+}
