+++ crates/system/os/accounts/src/sargon_os_sync_entities.rs
@@ -199,58 +199,6 @@ impl OsSyncEntitiesStateOnLedger for SargonOS {
     }
 }
 
-/// An action that implies a synchronisation needs to be applied in an entity in the local profile.
-#[derive(Clone, Debug, PartialEq, Eq, std::hash::Hash)]
-enum EntitySyncAction {
-    ToTombstone(AccountAddress),
-    ToSecurify(AddressOfAccountOrPersona, AccessControllerAddress),
-}
-
-/// The kinds of sync actions performed on entities in profile.
-#[derive(Clone, Debug, PartialEq, Eq, std::hash::Hash)]
-pub enum EntitySyncActionPerformed {
-    SomeEntitiesTombstoned,
-    SomeEntitiesSecurified,
-}
-
-impl HasSampleValues for EntitySyncActionPerformed {
-    fn sample() -> Self {
-        Self::SomeEntitiesTombstoned
-    }
-
-    fn sample_other() -> Self {
-        Self::SomeEntitiesSecurified
-    }
-}
-
-/// The report that gathers the different actions performed on profile after sync completes.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct EntitySyncReport {
-    pub actions_performed: IndexSet<EntitySyncActionPerformed>,
-}
-
-impl HasSampleValues for EntitySyncReport {
-    fn sample() -> Self {
-        Self::new(IndexSet::just(EntitySyncActionPerformed::sample()))
-    }
-
-    fn sample_other() -> Self {
-        Self::no_action()
-    }
-}
-
-impl EntitySyncReport {
-    pub fn new(actions: IndexSet<EntitySyncActionPerformed>) -> Self {
-        Self {
-            actions_performed: actions,
-        }
-    }
-
-    pub fn no_action() -> Self {
-        Self::new(IndexSet::new())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
