+++ crates/system/os/accounts/src/support/mod.rs
@@ -1,0 +1,3 @@
+mod entity_sync_report;
+
+pub use entity_sync_report::*;
